requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
google-generativeai==0.3.1
groq==0.4.1
requests==2.31.0
reportlab==4.0.7
Pillow==10.1.0
python-dotenv==1.0.0

